pyproject.toml
[project]
name = "xxx"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.12.0",
    "alembic>=1.18.4",
    "bitsandbytes>=0.49.2",
    "fastapi>=0.129.2",
    "openai>=2.24.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "qdrant-client>=1.17.0",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.46",
    "torch>=2.10.0",
    "transformers>=5.2.0",
    "uvicorn>=0.41.0",
]
